adding HomogenousAPITest
