+++ core/src/main/kotlin/com/strumenta/kolasu/model/Model.kt
@@ -1,181 +1,10 @@
 package com.strumenta.kolasu.model
 
-import com.strumenta.kolasu.model.annotations.Annotation
-import com.strumenta.kolasu.model.observable.Observer
-import java.io.Serializable
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
 import kotlin.reflect.KVisibility
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.memberProperties
-
-/**
- * The Abstract Syntax Tree will be constituted by instances of Node.
- */
-open class Node() : Serializable {
-
-    @Internal
-    private val annotations: MutableList<Annotation> = mutableListOf()
-
-    @Internal
-    val destinations = mutableListOf<Destination>()
-
-    constructor(range: Range?) : this() {
-        this.range = range
-    }
-
-    constructor(origin: Origin?) : this() {
-        this.origin = origin
-    }
-
-    @Internal
-    open val nodeType: String
-        get() = this::class.qualifiedName!!
-
-    /**
-     * The properties of this AST nodes, including attributes, children, and references.
-     */
-    @property:Internal
-    open val properties: List<PropertyDescription>
-        get() = try {
-            nodeProperties.map { PropertyDescription.buildFor(it, this) }
-        } catch (e: Throwable) {
-            throw RuntimeException("Issue while getting properties of node ${this::class.qualifiedName}", e)
-        }.also { properties ->
-            val alreadyFound = mutableSetOf<String>()
-            properties.forEach { property ->
-                val name = property.name
-                if (alreadyFound.contains(name)) {
-                    throw IllegalStateException("Duplicate property with name $name")
-                } else {
-                    alreadyFound.add(name)
-                }
-            }
-        }
-
-    /**
-     * The origin from which this AST Node has been generated, if any.
-     */
-    @property:Internal
-    var origin: Origin? = null
-
-    /**
-     * The parent node, if any.
-     */
-    @property:Internal
-    var parent: Node? = null
-
-    /**
-     * The range of this node in the source text.
-     * If a range has been provided when creating this node, it is returned.
-     * Otherwise, the value of this property is the range of the origin, if any.
-     */
-    @property:Internal
-    var range: Range?
-        get() = origin?.range
-        set(value) {
-            if (origin == null) {
-                if (value != null) {
-                    origin = SimpleOrigin(value)
-                }
-            } else {
-                origin!!.range = value
-            }
-        }
-
-    @property:Internal
-    val source: Source?
-        get() = origin?.source
-
-    /**
-     * Tests whether the given range is contained in the interval represented by this object.
-     * @param range the range
-     */
-    fun contains(range: Range?): Boolean {
-        return this.range?.contains(range) ?: false
-    }
-
-    /**
-     * Tests whether the given range overlaps the interval represented by this object.
-     * @param range the range
-     */
-    fun overlaps(range: Range?): Boolean {
-        return this.range?.overlaps(range) ?: false
-    }
-
-    /**
-     * The source text for this node
-     */
-    @Internal
-    val sourceText: String?
-        get() = origin?.sourceText
-
-    /**
-     * This must be final because otherwise data classes extending this will automatically generate
-     * their own implementation. If Link properties are present it could lead to stack overflows in case
-     * of circular graphs.
-     */
-    final override fun toString(): String {
-        return "${this.nodeType}(${properties.joinToString(", ") { "${it.name}=${it.valueToString()}" }})"
-    }
-
-    @property:Internal
-    val observers: MutableList<Observer<in Node>> = mutableListOf()
-    fun registerObserver(observer: Observer<*>) {
-        observers.add(observer as Observer<in Node>)
-    }
-
-    fun unregisterObserver(observer: Observer<in Node>) {
-        observers.remove(observer)
-    }
-
-    protected fun notifyOfPropertyChange(propertyName: String, oldValue: Any?, newValue: Any?) {
-        observers.forEach {
-            it.receivePropertyChangeNotification(this, propertyName, oldValue, newValue)
-        }
-    }
-
-    @Internal
-    val allAnnotations: List<Annotation>
-        get() = annotations
-
-    fun <I : Annotation>annotationsByType(kClass: KClass<I>): List<I> {
-        return annotations.filterIsInstance(kClass.java)
-    }
-
-    fun <I : Annotation>getSingleAnnotation(kClass: KClass<I>): I? {
-        val instances = annotations.filterIsInstance(kClass.java)
-        return if (instances.isEmpty()) {
-            null
-        } else if (instances.size == 1) {
-            instances.first()
-        } else {
-            throw IllegalStateException("More than one instance of $kClass found")
-        }
-    }
-
-    fun <A : Annotation>addAnnotation(annotation: A): A {
-        if (annotation.annotatedNode != null) {
-            throw java.lang.IllegalStateException("Annotation already attached")
-        }
-        annotation.attachTo(this)
-        if (annotation.single) {
-            annotations.filter { it.annotationType == annotation.annotationType }.forEach { removeAnnotation(it) }
-        }
-        annotations.add(annotation)
-        return annotation
-    }
-
-    fun removeAnnotation(annotation: Annotation) {
-        require(annotation.annotatedNode == this)
-        annotations.remove(annotation)
-        annotation.detach()
-    }
-
-    fun hasAnnotation(annotation: Annotation): Boolean {
-        return annotations.contains(annotation)
-    }
-}
 
 fun <N : Node> N.withRange(range: Range?): N {
     this.range = range
